relocate init_checkpoints as they are only used in tests
